Bug Fixes
fixed some faulty logic that caused data to go out of whack when building the octree. Now all the data in the octree seems to be where it belongs. The tree is being built using the temp.ptx but the rabbit also works just. Note increasing the  presetBinLevel to ~3 adds levels to the tree which is required for the larger files. We can probably add a check to set this automatically by looking at the number of points in the file and choosing the closest power of 8